MapModule as ES module
